requirements.txt
Django==1.10
djangorestframework==3.7.7
MySQL-python==1.2.5

